:rocket: Deploy app

+++ src/components/Collections/Personal.js
@@ -19,7 +19,6 @@ import Todo from "../Todo/Todo"
 import Navbar from "../Navbar/Navbar"
 import SideNav from "../SideNav/SideNav"
 import CompletedTodo from "../Todo/CompletedTodo"
-import { UserAuth } from "../../HOC/AuthContext"
 import './collection.scss'
 
 const Personal = () => {
@@ -32,8 +31,7 @@ const Personal = () => {
     
 
     const router = useNavigate()
-    //const user = auth.currentUser
-    const { user } = UserAuth()
+    const user = auth.currentUser
 
     useEffect(() => {
         getUncompleteTasks()
@@ -41,45 +39,50 @@ const Personal = () => {
     }, [])
 
     const addTodo = async () => {
-        //if (user !== null) {
-            //const uid = user.uid
-            const collectionRef = collection(db, `personal/${user?.uid}/todoList`)
-            const payload = {
-                todo,
-                //creates a timestamp which is unique so we use this as the key when returning documents in the subcollection.
-                time: serverTimestamp(),
-                complete: false
+        if (user !== null) {
+            const uid = user.uid
+            if(todo !== '') {
+                const collectionRef = collection(db, `personal/${uid}/todoList`)
+                const payload = {
+                    todo,
+                    //creates a timestamp which is unique so we use this as the key when returning documents in the subcollection.
+                    time: serverTimestamp(),
+                    complete: false
+                }
+                await addDoc(collectionRef, payload)
+            } else {
+                setTodo('')
             }
-            await addDoc(collectionRef, payload)
-        //}
+        }
+        setTodo('')
     }
 
     const deleteTodo = async (id) => {
-        //if (user !== null ){
+        if (user !== null ){
             //fetches the user's uid
-            //const uid = user.uid
-            const docRef = doc(db, `/personal/${user?.uid}/todoList`, id)
+            const uid = user.uid
+            const docRef = doc(db, `/personal/${uid}/todoList`, id)
             await deleteDoc(docRef)
-        //}
+        }
     }
 
     const editTodo = async (id, todo) => {
-        //if (user !== null) {
-            // const uid = user.uid
-            const todoRef = doc(db, `personal/${user?.uid}/todoList/${id}`)
+        if (user !== null) {
+            const uid = user.uid
+            const todoRef = doc(db, `personal/${uid}/todoList/${id}`)
             await setDoc (todoRef, {
                 todo: todo
             }, {merge: true})
-        //}
+        }
     }
 
     const getUncompleteTasks = () => {
-        //if (user !== null ){
+        if (user !== null ){
             //fetches the user's uid
-            // const uid = user.uid
+            const uid = user.uid
             //uses the uid as the document id in school collection and then creates a subcollection called todoList
             //returns an array of documents with "complete: false"
-            const q = query(collection(db, `personal/${user?.uid}/todoList`), where('complete', '==', false), orderBy('time', 'desc'))
+            const q = query(collection(db, `personal/${uid}/todoList`), where('complete', '==', false), orderBy('time', 'desc'))
             onSnapshot(q, (querySnapshot) => {
                 let items = []
                 querySnapshot.docs.map((doc) => {
@@ -90,16 +93,16 @@ const Personal = () => {
                 setUncompletedTasks(items)
                 setNumOfUncomplete(items.length)
             })
-        //}
+        }
     }
 
     const getCompleteTasks = () => {
-        // if (user !== null ){
+        if (user !== null ){
             //fetches the user's uid
-            // const uid = user.uid
+            const uid = user.uid
             //uses the uid as the document id in school collection and then creates a subcollection called todoList
             //returns an array of documents with "complete: true"
-            const q = query(collection(db, `personal/${user?.uid}/todoList`), where('complete', '==', true), orderBy('time', 'desc'))
+            const q = query(collection(db, `personal/${uid}/todoList`), where('complete', '==', true), orderBy('time', 'desc'))
             onSnapshot(q, (querySnapshot) => {
                 let items = []
                 querySnapshot.docs.map((doc) => {
@@ -110,30 +113,30 @@ const Personal = () => {
                 setCompletedTasks(items)
                 setNumOfComplete(items.length)
             })
-        //}
+        }
     }
 
     const checkComplete = async (id) => {
-        // if (user !== null ){
-            // const uid = user.uid
-            const docRef = doc(db, `/personal/${user?.uid}/todoList`, id)
+        if (user !== null ){
+            const uid = user.uid
+            const docRef = doc(db, `/personal/${uid}/todoList`, id)
             const payload = {
                 complete : true,
                 time: serverTimestamp()
             }
             await setDoc(docRef, payload, {merge: true})
-        //}
+        }
     }
 
     const checkUncomplete = async (id) => {
-        // if (user !== null ){
-            // const uid = user.uid
-            const docRef = doc(db, `/personal/${user?.uid}/todoList`, id)
+        if (user !== null ){
+            const uid = user.uid
+            const docRef = doc(db, `/personal/${uid}/todoList`, id)
             const payload = {
                 complete : false
             }
             await setDoc(docRef, payload, {merge: true})
-        //}
+        }
     }
 
     return (
@@ -155,7 +158,7 @@ const Personal = () => {
                             </button>
                         </div>
                         <div className="todo-form">
-                            <button className='add-btn' onClick={() => {addTodo(); setTodo('')}}><HiPlusSm /></button>
+                            <button className='add-btn' onClick={addTodo}><HiPlusSm /></button>
                             <input 
                                 type='text' 
                                 className='input'
